+++ boofuzz/connections/serial_connection.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 
 import time
 
-from . import itarget_connection, serial_connection_low_level
+from boofuzz.connections import itarget_connection, serial_connection_low_level
 
 
 class SerialConnection(itarget_connection.ITargetConnection):
@@ -30,6 +30,10 @@ class SerialConnection(itarget_connection.ITargetConnection):
                return 0
 
     If none of these methods are used, your connection may hang forever.
+
+    .. versionchanged:: 0.2.0
+        SerialConnection has been moved into the connections subpackage.
+        The full path is now boofuzz.connections.serial_connection.SerialConnection
 
     Args:
         port (Union[int, str]): Serial port name or number.