+++ boofuzz/utils/dcerpc.py
@@ -18,28 +18,28 @@ def bind(uuid, version):
     major = struct.pack("<H", int(major))
     minor = struct.pack("<H", int(minor))
 
-    bind_request = "\x05\x00"  # version 5.0
-    bind_request += "\x0b"  # packet type = bind (11)
-    bind_request += "\x03"  # packet flags = last/first flag set
-    bind_request += "\x10\x00\x00\x00"  # data representation
-    bind_request += "\x48\x00"  # frag length: 72
-    bind_request += "\x00\x00"  # auth length
-    bind_request += "\x00\x00\x00\x00"  # call id
-    bind_request += "\xb8\x10"  # max transmit frag (4280)
-    bind_request += "\xb8\x10"  # max recv frag (4280)
-    bind_request += "\x00\x00\x00\x00"  # assoc group
-    bind_request += "\x01"  # number of ctx items (1)
-    bind_request += "\x00\x00\x00"  # padding
-    bind_request += "\x00\x00"  # context id (0)
-    bind_request += "\x01"  # number of trans items (1)
-    bind_request += "\x00"  # padding
+    bind_request = b"\x05\x00"  # version 5.0
+    bind_request += b"\x0b"  # packet type = bind (11)
+    bind_request += b"\x03"  # packet flags = last/first flag set
+    bind_request += b"\x10\x00\x00\x00"  # data representation
+    bind_request += b"\x48\x00"  # frag length: 72
+    bind_request += b"\x00\x00"  # auth length
+    bind_request += b"\x00\x00\x00\x00"  # call id
+    bind_request += b"\xb8\x10"  # max transmit frag (4280)
+    bind_request += b"\xb8\x10"  # max recv frag (4280)
+    bind_request += b"\x00\x00\x00\x00"  # assoc group
+    bind_request += b"\x01"  # number of ctx items (1)
+    bind_request += b"\x00\x00\x00"  # padding
+    bind_request += b"\x00\x00"  # context id (0)
+    bind_request += b"\x01"  # number of trans items (1)
+    bind_request += b"\x00"  # padding
     bind_request += helpers.uuid_str_to_bin(uuid)  # abstract syntax
     bind_request += major  # interface version
     bind_request += minor  # interface version minor
 
     # transfer syntax 8a885d04-1ceb-11c9-9fe8-08002b104860 v2.0
-    bind_request += "\x04\x5d\x88\x8a\xeb\x1c\xc9\x11\x9f\xe8\x08\x00\x2b\x10\x48\x60"
-    bind_request += "\x02\x00\x00\x00"
+    bind_request += b"\x04\x5d\x88\x8a\xeb\x1c\xc9\x11\x9f\xe8\x08\x00\x2b\x10\x48\x60"
+    bind_request += b"\x02\x00\x00\x00"
 
     return bind_request
 
@@ -50,11 +50,11 @@ def bind_ack(data):
     """
 
     # packet type == bind ack (12)?
-    if data[2] != "\x0c":
+    if data[2] != b"\x0c":
         return False
 
     # ack result == acceptance?
-    if data[36:38] != "\x00\x00":
+    if data[36:38] != b"\x00\x00":
         return False
 
     return True
@@ -81,19 +81,19 @@ def request(opnum, data):
         if i == num_frags - 1:
             flags |= 0x2  # last frag
 
-        request_buffer = "\x05\x00"  # version 5.0
-        request_buffer += "\x00"  # packet type = request (0)
+        request_buffer = b"\x05\x00"  # version 5.0
+        request_buffer += b"\x00"  # packet type = request (0)
         request_buffer += struct.pack("B", flags)  # packet flags
-        request_buffer += "\x10\x00\x00\x00"  # data representation
+        request_buffer += b"\x10\x00\x00\x00"  # data representation
         request_buffer += frag_length  # frag length
-        request_buffer += "\x00\x00"  # auth length
-        request_buffer += "\x00\x00\x00\x00"  # call id
+        request_buffer += b"\x00\x00"  # auth length
+        request_buffer += b"\x00\x00\x00\x00"  # call id
         request_buffer += alloc_hint  # alloc hint
-        request_buffer += "\x00\x00"  # context id (0)
+        request_buffer += b"\x00\x00"  # context id (0)
         request_buffer += struct.pack("<H", opnum)  # opnum
         request_buffer += chunk
 
         frags.append(request_buffer)
 
     # you don't have to send chunks out individually. so make life easier for the user and send them all at once.
-    return "".join(frags)
+    return b"".join(frags)